+++ src/Query/Builder.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types = 1);
+
+namespace Formal\AccessLayer\Query;
+
+use Formal\AccessLayer\Query;
+
+/**
+ * @psalm-immutable
+ */
+interface Builder
+{
+    public function normalize(): Query;
+}